requirements.txt
dj-database-url==0.4.0
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
Django==1.9.7
django-oauth-toolkit==0.10.0
django-rest-swagger==0.3.6
djangorestframework==3.3.3
ipython==4.2.0
django-ckeditor==5.0.3
python-slugify==1.2.0
drfdocs